Develop

+++ apps/api/src/modules/user/user.service.ts
@@ -1,18 +1,15 @@
-import {
-  BadRequestException,
-  Inject,
-  Injectable,
-  NotFoundException,
-} from '@nestjs/common';
-import * as schema from '~/database/schema';
-import { NodePgDatabase } from 'drizzle-orm/node-postgres';
-import { DrizzleAsyncProvider } from '~/database/drizzle.provider';
-import { UserResponseType } from '@family-tree/shared';
-import { and, eq, isNull } from 'drizzle-orm';
-import { UserUpdateRequestDto } from './dto/user.dto';
-import { CloudflareConfig } from '~/config/cloudflare/cloudflare.config';
-import { EnvType } from '~/config/env/env-validation';
+import { randomUUID } from 'node:crypto';
+import type { UserResponseType } from '@family-tree/shared';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
+import { and, eq, isNull } from 'drizzle-orm';
+import { NodePgDatabase } from 'drizzle-orm/node-postgres';
+import { CloudflareConfig } from '~/config/cloudflare/cloudflare.config';
+import type { EnvType } from '~/config/env/env-validation';
+import { DrizzleAsyncProvider } from '~/database/drizzle.provider';
+import * as schema from '~/database/schema';
+import { DICEBAR_URL } from '~/utils/constants';
+import { UserUpdateRequestDto } from './dto/user.dto';
 
 @Injectable()
 export class UserService {
@@ -22,7 +19,7 @@ export class UserService {
     @Inject(DrizzleAsyncProvider)
     private db: NodePgDatabase<typeof schema>,
     private cloudflareConfig: CloudflareConfig,
-    private configService: ConfigService<EnvType>
+    configService: ConfigService<EnvType>,
   ) {
     this.cloudflareR2Path =
       configService.getOrThrow<EnvType['CLOUDFLARE_URL']>('CLOUDFLARE_URL');
@@ -32,7 +29,7 @@ export class UserService {
     const user = await this.db.query.usersSchema.findFirst({
       where: and(
         eq(schema.usersSchema.email, email),
-        isNull(schema.usersSchema.deletedAt)
+        isNull(schema.usersSchema.deletedAt),
       ),
     });
 
@@ -47,7 +44,7 @@ export class UserService {
     const user = await this.db.query.usersSchema.findFirst({
       where: and(
         eq(schema.usersSchema.id, id),
-        isNull(schema.usersSchema.deletedAt)
+        isNull(schema.usersSchema.deletedAt),
       ),
     });
 
@@ -62,7 +59,7 @@ export class UserService {
     const user = await this.db.query.usersSchema.findFirst({
       where: and(
         eq(schema.usersSchema.id, id),
-        isNull(schema.usersSchema.deletedAt)
+        isNull(schema.usersSchema.deletedAt),
       ),
     });
 
@@ -77,7 +74,7 @@ export class UserService {
     const user = await this.db.query.usersSchema.findFirst({
       where: and(
         eq(schema.usersSchema.id, id),
-        isNull(schema.usersSchema.deletedAt)
+        isNull(schema.usersSchema.deletedAt),
       ),
     });
 
@@ -85,15 +82,15 @@ export class UserService {
       throw new NotFoundException(`User with id ${id} not found`);
     }
 
-    if (!user.image?.includes(this.cloudflareR2Path)) {
-      throw new BadRequestException('Image is not uploaded');
-    }
-
     if (user.gender !== body.gender) {
       // FIXME: Need to think about related family trees
     }
 
-    if (user.image && user.image !== body.image) {
+    if (
+      user.image &&
+      user.image !== body.image &&
+      user.image?.includes(this.cloudflareR2Path)
+    ) {
       this.cloudflareConfig.deleteFile(user.image);
     }
 
@@ -103,5 +100,32 @@ export class UserService {
         ...body,
       })
       .where(eq(schema.usersSchema.id, id));
+  }
+
+  async updateUserAvatar(id: string): Promise<UserResponseType> {
+    const user = await this.db.query.usersSchema.findFirst({
+      where: and(
+        eq(schema.usersSchema.id, id),
+        isNull(schema.usersSchema.deletedAt),
+      ),
+    });
+
+    if (!user) {
+      throw new NotFoundException(`User with id ${id} not found`);
+    }
+
+    if (user.image?.includes(this.cloudflareR2Path)) {
+      this.cloudflareConfig.deleteFile(user.image);
+    }
+
+    const [updatedUser] = await this.db
+      .update(schema.usersSchema)
+      .set({
+        image: `${DICEBAR_URL}/7.x/notionists/svg?seed=${randomUUID()}`,
+      })
+      .where(eq(schema.usersSchema.id, id))
+      .returning();
+
+    return updatedUser;
   }
 }
